+++ NEWS.md
@@ -1,5 +1,7 @@
 # cpp11 (development version)
 
+* Fixed a small protection issue flagged by rchk (#408).
+
 * `R_NO_REMAP` and `STRICT_R_HEADERS` are now conditionally defined only if they
   have not already been defined elsewhere. This is motivated by the fact that
   `R_NO_REMAP` is becoming the default for C++ code in R 4.5.0 (#410).